+++ tests/test_encoding.py
@@ -16,36 +16,43 @@ class EncodingTest(unittest.TestCase):
         self.assertEqual(uridecode(encoded.decode('ascii'), encoding), decoded)
 
     def test_encoding(self):
-        for decoded, encoded in [
+        cases = [
             ('', b''),
-            ('~guido', b'~guido'),
+            (' ', b'%20'),
+            ('%', b'%25'),
+            ('~', b'~'),
             (UNRESERVED.decode('ascii'), UNRESERVED),
-        ]:
+        ]
+        for decoded, encoded in cases:
             self.check(decoded, encoded)
 
     def test_safe_encoding(self):
-        for decoded, encoded, safe in [
+        cases = [
             ('', b'', b''),
-            (UNRESERVED.decode('ascii'), UNRESERVED, b''),
-            (UNRESERVED.decode('ascii'), UNRESERVED, UNRESERVED),
+            (' ', b' ', b' '),
+            ('%', b'%', b'%'),
             (RESERVED.decode('ascii'), RESERVED, RESERVED)
-        ]:
+        ]
+        for decoded, encoded, safe in cases:
             self.check(decoded, encoded, safe)
 
     def test_utf8_encoding(self):
-        for decoded, encoded in [
+        cases = [
             ('\xf6lk\xfcrbis', b'%C3%B6lk%C3%BCrbis')
-        ]:
+        ]
+        for decoded, encoded in cases:
             self.check(decoded, encoded, encoding='utf-8')
 
     def test_latin1_encoding(self):
-        for decoded, encoded in [
+        cases = [
             ('\xf6lk\xfcrbis', b'%F6lk%FCrbis')
-        ]:
+        ]
+        for decoded, encoded in cases:
             self.check(decoded, encoded, encoding='latin-1')
 
     def test_idna_encoding(self):
-        for decoded, encoded in [
+        cases = [
             ('\xf6lk\xfcrbis', b'xn--lkrbis-vxa4c')
-        ]:
+        ]
+        for decoded, encoded in cases:
             self.check(decoded, encoded, encoding='idna')